Here goes Beam Search implementation.
However, it messes up the search due to something we did not think of before. Current test problem generates 360 neighbors for each solution, then, setting the beam length lower than 360 makes it act exactly like naive hill climbing because only the neighbors of best agent are left within the beam after first iteration, then guess what? It continues to expand the neighbors of that so similar solutions derived from first best. Setting a beam length like 10000 would work though... If you can bring back a computer from 2050 because then, it is exactly like breadth first search.

This might work on the oop design problem as the neighbor count will be much smaller (I hope, anyway). Still, to be on the safe side, we should modify it so that each agent maintains it's own beam or something like that. Maybe omit this entirely or just keep it to show that it's worst method in this domain.